Consolidate biography cleanup into shared helper function
